requirements.txt
fastapi
uvicorn
gunicorn
pandas
python-dotenv
langchain
langchain-pinecone
langchain-groq
pinecone-client